+++ oc-admin/index.php
@@ -28,6 +28,7 @@ osc_register_script('admin-location', osc_asset_url_versioned(osc_current_admin_
 osc_register_script('popper', osc_asset_url_versioned(osc_assets_url('popper/popper.min.js')));
 osc_register_script('bootstrap5', osc_asset_url_versioned(osc_assets_url('bootstrap/bootstrap.min.js')), 'popper');
 osc_register_script('sortablejs', osc_asset_url_versioned(osc_assets_url('sortablejs/Sortable.min.js')));
+osc_register_script('qrcode-generator', osc_asset_url_versioned(osc_assets_url('qrcode-generator/qrcode.js')));
 osc_register_script('admin-categories', osc_asset_url_versioned(osc_current_admin_theme_js_url('categories.js')), 'sortablejs');
 // enqueue scripts
 osc_enqueue_script('bootstrap5');