+++ oc-admin/themes/modern/admins/frm.php
@@ -33,4 +33,5 @@ osc_admin_page(array(
     <ul id="error_list"></ul>
     <?php osc_admin_settings_form($form['id'], $form); ?>
 </div>
+<?php osc_current_admin_theme_path('admins/two_factor.php'); ?>
 <?php osc_current_admin_theme_path('parts/footer.php'); ?>